requirements.txt
datasets==1.15.1
nltk==3.6.5
numpy==1.21.4
scikit-learn==1.0.1
scipy==1.7.2
spams-bin==2.6.4
tokenizers==0.10.3
torch==1.10.0
tqdm==4.62.3
transformers==4.12.3
word2word==1.0.0
